added resource routes

+++ app/routes/_cms+/admin.content.$id.$entryId.tsx
@@ -1,7 +1,9 @@
 import type { ActionArgs, LoaderArgs } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
-import { Form, useLoaderData } from "@remix-run/react";
+import { Form, useLoaderData, useSubmit } from "@remix-run/react";
+import { useRef } from "react";
 import invariant from "tiny-invariant";
+import { DynamicField } from "~/cms/components/forms/fields";
 import { RouteTitle } from "~/cms/components/route-title";
 import { Button } from "~/cms/components/ui/button";
 import { deleteEntry, getEntryById } from "~/models/content.server";
@@ -16,38 +18,58 @@ export const loader = async ({ params }: LoaderArgs) => {
   return json({ entry });
 };
 
-export const action = async ({ params }: ActionArgs) => {
+export const action = async ({ params, request }: ActionArgs) => {
   invariant(params.entryId, "Missing entry id");
-  try {
-    await deleteEntry(params.entryId);
-    return redirect(`/admin/content/${params.id}`);
-  } catch (error) {
-    console.error(error);
-    throw "Error deleting entry";
+  const formData = await request.formData();
+
+  if (request.method === "POST") {
+    // update entry field values
+    console.log(Object.fromEntries(formData.entries()));
+    return { success: true };
   }
+
+  if (request.method === "DELETE") {
+    try {
+      await deleteEntry(params.entryId);
+      return redirect(`/admin/content/${params.id}`);
+    } catch (error) {
+      console.error(error);
+      throw "Error deleting entry";
+    }
+  }
+
+  return { success: false };
 };
 
 export default function EntryEditRoute() {
   const { entry } = useLoaderData<typeof loader>();
 
+  const submit = useSubmit();
+  const updateEntryForm = useRef(null);
+
+  function updateEntry() {
+    submit(updateEntryForm.current);
+  }
+
   return (
     <div>
       <RouteTitle title="Entry">
-        <Form method="delete">
+        <Form method="DELETE">
           <Button type="submit" variant="destructive">
             Delete
           </Button>
         </Form>
+        <Button onClick={updateEntry}>Save Changes</Button>
       </RouteTitle>
-      <ul>
-        {entry.fields
-          .sort((a, b) => a.field.sortOrder - b.field.sortOrder)
-          .map((field) => (
-            <li key={field.id}>
-              {field.field.title} - {field.field.type} - {field.field.sortOrder}
-            </li>
-          ))}
-      </ul>
+      <Form
+        ref={updateEntryForm}
+        method="POST"
+        className="max-w-5xl grid gap-4 pb-8"
+      >
+        {entry.fields.map((field) => (
+          <DynamicField key={field.id} field={field.field} values={field} />
+        ))}
+      </Form>
     </div>
   );
 }
